+++ demo/index.js
@@ -50,6 +50,8 @@ if (window.innerWidth < 768) {
 }
 
 window.vditor = new Vditor('vditor', {
+  // _lutePath: `http://192.168.0.107:9090/lute.min.js?${new Date().getTime()}`,
+  _lutePath: 'src/js/lute/lute.min.js',
   toolbar,
   mode: 'sv',
   height: window.innerHeight + 100,
@@ -77,11 +79,14 @@ window.vditor = new Vditor('vditor', {
       'j': 'https://unpkg.com/vditor@1.3.1/dist/images/emoji/j.png',
     },
     at: (key) => {
-      return [
-        {
-          value: '@Vanessa',
-          html: '<img src="https://avatars0.githubusercontent.com/u/970828?s=60&v=4"/> Vanessa',
-        }]
+      if ('vanessa'.indexOf(key.toLocaleLowerCase()) > -1) {
+        return [
+          {
+            value: '@Vanessa',
+            html: '<img src="https://avatars0.githubusercontent.com/u/970828?s=60&v=4"/> Vanessa',
+          }]
+      }
+      return []
     },
   },
   tab: '\t',